feat: moved alerts and simulation sections on separate folders (PIN-9718)
